voicebridge-backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
openai-whisper
# CPU-only PyTorch (prevents WinError 1114)
torch==2.3.0+cpu
torchaudio==2.3.0+cpu
--extra-index-url https://download.pytorch.org/whl/cpu

numpy==1.26.4
requests>=2.28.0
transformers>=4.30.0
huggingface-hub>=0.15.0
datasets>=2.12.0
soundfile>=0.12.0
librosa>=0.10.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-magic>=0.4.27
aiofiles>=23.0.0
accelerate>=0.20.0
